Adjust the availability of buttons in the user editor

+++ music_paradigm_web/src/components/admin/users/widget-editor/users-editor.widget.vue
@@ -1,20 +1,20 @@
 <template>
 	<div id="users-editor" class="widget widget-bg">
 		<div class="edition-buttons-position">
-			<button v-on:click="handleRevert" class="widget-button blue">Revert</button>
-			<button v-on:click="handleUnselection" class="widget-button blue">Unselect</button>
+			<button v-on:click="handleRevert" class="widget-button blue" :class="{ inactive: !isRevertButtonActive }">Revert</button>
+			<button v-on:click="handleUnselection" class="widget-button blue" :class="{ inactive: !isUnselectButtonActive }">Unselect</button>
 		</div>
 
 		<users-editor-form-component class="editor-position inner-widget" ref="editorForm" />
 
 		<div class="submission-buttons-position">
-			<form submit.prevent ref="upload" style="display: none">
+			<!-- <form submit.prevent ref="upload" style="display: none">
 				<input type="file" id="myfile" name="myfile" v-on:change="handleUploadUsers" ref="fileInput" />
 			</form>
-			<button class="widget-button inactive">Upload</button>
+			<button class="widget-button inactive">Upload</button> -->
 			<button v-on:click="submitUserToCreate" class="widget-button green">Create</button>
-			<button v-on:click="submitUserToUpdate" class="widget-button blue">Update</button>
-			<button v-on:click="submitUserToDelete" class="widget-button red">Delete</button>
+			<button v-on:click="submitUserToUpdate" class="widget-button blue" :class="{ inactive: !isUpdateButtonActive }">Update</button>
+			<button v-on:click="submitUserToDelete" class="widget-button red" :class="{ inactive: !isDeleteButtonActive }">Delete</button>
 		</div>
 	</div>
 </template>
@@ -30,9 +30,21 @@ export default {
 	components: {
 		UsersEditorFormComponent,
 	},
+	data() {
+		return {
+			isRevertButtonActive: false,
+			isUpdateButtonActive: false,
+		};
+	},
 	computed: {
 		...mapGetters('curriculums', ['curriculumsHeadersList']),
 		...mapGetters('users', ['hasSelectedUser', 'userSelectedId']),
+		isUnselectButtonActive() {
+			return this.hasSelectedUser;
+		},
+		isDeleteButtonActive() {
+			return this.hasSelectedUser;
+		},
 	},
 	methods: {
 		...mapActions('curriculums', ['fetchAllCurriculumHeaders']),
@@ -47,6 +59,7 @@ export default {
 			this.$emit('create-user');
 		},
 		submitUserToUpdate() {
+			if (!this.isUpdateButtonActive) return;
 			const answer = window.confirm('Are your sure you want to edit the user(s)?');
 			if (answer) {
 				const userToCreate = this.bundleUserFromForm();
@@ -57,18 +70,28 @@ export default {
 			}
 		},
 		submitUserToDelete() {
+			if (!this.isDeleteButtonActive) return;
 			const answer = window.confirm('Are your sure you want to delete the user(s)?');
 			if (answer) this.deleteUser(this.userSelectedId);
 		},
 		handleRevert() {
+			if (!this.isRevertButtonActive) return;
 			this.assignSelectedToForm();
 		},
 		handleUnselection() {
+			if (!this.isUnselectButtonActive) return;
 			this.unsetSelectedUser();
 		},
 		handleUploadUsers() {
 			console.log('To implement');
 		},
+		evaluateIsRevertButtonActive(wasFormModified) {
+			this.isRevertButtonActive = wasFormModified;
+			this.isUpdateButtonActive = this.hasSelectedUser && wasFormModified;
+		},
+	},
+	mounted() {
+		this.$watch(() => this.$refs.editorForm.wasFormModified, this.evaluateIsRevertButtonActive, { immediate: true });
 	},
 };
 </script>
@@ -89,7 +112,7 @@ export default {
 	grid-area: submission-btn;
 	display: grid;
 	grid-gap: 15px;
-	grid-template-columns: 1fr 1fr 1fr 1fr;
+	grid-template-columns: 1fr 1fr 1fr;
 }
 
 .widget {
